Automatic Ticket Classfication tool/requirements.txt
gradio
langchain
langchain-openai
langchain-community
langchain-pinecone
openai
tiktoken
python-dotenv
pinecone-client
pypdf
joblib
pandas
scikit-learn
sentence-transformers